fixup! feat(bazel/integration): support tracking sizes within the integration test rule directly
